requirements.txt
# MCP server (existing)
mcp>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
gspread>=6.0.0
google-auth>=2.29.0
python-dotenv>=1.0.0

# Web dashboard backend
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
supabase>=2.0.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0
